Proposed extension
Added a couple of clarifications to the paragraph on the induction axiom (this name wasn't introduced).
Added a remark and associated exercise problem: axioms for less-than, addition, and multiplication aren't needed in PA2, since these are definable.
(Too much to leave both as exercises?  Should we prove it for < and + and leave multiplication as a exercise? Or is all this too much?)