Some corrections and proposal for SOL (#139)
* variables -> object variables &minor typos

* Typo, (bound variable) u -> X

* Typo, V_i^n -> u_i^n

* Proposed extension

Added a couple of clarifications to the paragraph on the induction axiom (this name wasn't introduced).
Added a remark and associated exercise problem: axioms for less-than, addition, and multiplication aren't needed in PA2, since these are definable.
(Too much to leave both as exercises?  Should we prove it for < and + and leave multiplication as a exercise? Or is all this too much?)

* Missing quantifiers in cor:sol-not-axiomatizable

Since we defines P as the conjunctions of all nine axioms, + and \times have to be generalized away as well.
Alternatively, we could use PA2\dagger (as newly introduced above: only successor axioms and induction) and simplify the proof here.

* \dom -> \ran & qualified "infinite"

I turned "infinite" into "(Dedekind) infinite".  Without explanation that's perhaps a bit obscure? Leave it? Explain it?

* Update introduction.tex