Quite a few typos in The Power of the Continuum.
There's quite a few errors here that I think I've corrected properly,
I'm not sure what's going on here. In detail:

* Definition 7.34. It makes no sense that x, a member of Y, R-codes Y,
nor that Y is a member of Z, the subset of X, whose powerset we want to
code with Y. I'm guessing what was meant to be said was that every
subset Z of X is R-coded by some element x in Y, and every element x in
Y R-codes some subset Z of X.

* Proposition 7.36. First off, the arity of Pow is incorrect here: there
should be three arguments. Following the style of the previous
proposition, I'm guessing it's meant to be Pow(Y, R, X), which means
that it's also missing a forall Y quantifier at the beginning. Then we
can remove the forall Y quantifier later down. In words, the fixed
sentence would be expressing that if Y R-codes the powerset of X, then
there is no injective function from Y to X; that is, ¬(℘(X) ⪯ X). (It
would have been very helpful to have this statement in words first.)

* Surely the power of the continuum is the size of the powerset of N,
not the powerset of R, if it's to have the same size as R.

* Proposition 7.37. Cont(X) suggests that X is a free variable in this
formula, but the formula immediately binds X, Y, R in foralls. Since we
want the free variable to be interpreted as the set with the same size
as R, it should be Y that is free, since it R-codes the powerset of the
countably infinite set X. So the forall Y quantifier should be removed,
and the property the formula expresses is that s(Y) = R.

* Proposition 7.40. The continuum hypothesis is false if the cardinal
aleph-1 is strictly in between aleph-0 and the power of the continuum.
Taking Y to be the set with cardinality aleph-1, to be strictly greater
than aleph-0, we need Y to be uncountable. Therefore the middle conjunct
should be acting on Y, not X. X is trivially uncountable anyway, since
it has cardinality 2^aleph-0.