Some typos (#230)
* Minor typo: extra s

* Fixed typo

c < k should be c < 1÷k in the completeness example on infinitesimals.

* Minor typo

* Minor typo: missing a

* Minor typo: missing )

* Minor typo

* Minor typo: missing period.

* Minor typo: period to outside math env

* Minor typo: missing )

* Minor typo: more missing ))

* Minor typo: should be → instead of ∨

* Clarify that P, Q are p.r. in Prop 2.16.

* R should be P instead, I think? R isn't defined anywhere in Prop 2.16 and the proof of \lnot uses P instead of R.

* Fix proof of Prop 2.17: Quantifier bounded by z < y, so range should be 0, ..., y - 1, and we never need to consider R(x, y).

* Typos

* Multiple typos.

* Minor typo: missing )

* Some clarification edits. The second sentence of that first paragraph is still a bit long and confusing to parse with the two 'that's, but this is what I could come up with. I hope 'took it' was a typo for 'took it to mean', otherwise I have no idea what the intention was.

* Minor typo: missing y

* Minor typo

* Minor typos; incorrect index, most consistent wording, incorrect variable

* Minor typos

* Added missing free variable for clarity

* Minor typos: punctuation, capitalization, Löb's proves the second and not the first incompleteness theorem

* Minor typo

* Minor typos: Refer to part (b) instead of nonexistent (2), use a instead of x for clarity, spelling

* Minor typos: some formatting, missing period

* Minor typos: brackets, parens, spelling, \le instead of <

* Minor typos

* Quite a few typos in The Power of the Continuum.

There's quite a few errors here that I think I've corrected properly,
I'm not sure what's going on here. In detail:

* Definition 7.34. It makes no sense that x, a member of Y, R-codes Y,
nor that Y is a member of Z, the subset of X, whose powerset we want to
code with Y. I'm guessing what was meant to be said was that every
subset Z of X is R-coded by some element x in Y, and every element x in
Y R-codes some subset Z of X.

* Proposition 7.36. First off, the arity of Pow is incorrect here: there
should be three arguments. Following the style of the previous
proposition, I'm guessing it's meant to be Pow(Y, R, X), which means
that it's also missing a forall Y quantifier at the beginning. Then we
can remove the forall Y quantifier later down. In words, the fixed
sentence would be expressing that if Y R-codes the powerset of X, then
there is no injective function from Y to X; that is, ¬(℘(X) ⪯ X). (It
would have been very helpful to have this statement in words first.)

* Surely the power of the continuum is the size of the powerset of N,
not the powerset of R, if it's to have the same size as R.

* Proposition 7.37. Cont(X) suggests that X is a free variable in this
formula, but the formula immediately binds X, Y, R in foralls. Since we
want the free variable to be interpreted as the set with the same size
as R, it should be Y that is free, since it R-codes the powerset of the
countably infinite set X. So the forall Y quantifier should be removed,
and the property the formula expresses is that s(Y) = R.

* Proposition 7.40. The continuum hypothesis is false if the cardinal
aleph-1 is strictly in between aleph-0 and the power of the continuum.
Taking Y to be the set with cardinality aleph-1, to be strictly greater
than aleph-0, we need Y to be uncountable. Therefore the middle conjunct
should be acting on Y, not X. X is trivially uncountable anyway, since
it has cardinality 2^aleph-0.

* Fixed wording in Church-Turing thesis; other minor typos

* Minor fixes:

* Added a few more parentheses to the examples in the notational
conventions of lambda calculus, just to be extra clear in the
beginning here.
* The substitution notation used here looks like M[N/x], so that
notation isn't an "alternative". It was probably meant to be [x/N]M.