Possible typo in primitive-recursion.text
The current passage from lines 99 to 102 reads as follows:

The general pattern then is this: to give a primitive recursive definition of a function $h(x_0, \dotsc, x_{k-1}, y)$, we provide two equations. The first defines the value of  $h(x_0, \dotsc, x_{k-1}, y)$ without reference to $f$.

It seems like the reference to $f$ might be a typo or ambiguous, since $f$ is last mentioned on line 50. The passage seems to make more sense if it were to say that the initial value of $h$ is defined without reference to itself, instead of defined without reference to $f$.