Fixed "total order" definitions; fixed several typos.
Issue #334: definition 2.16, definition 2.24, and the footnote of section 5.6 provided conflicting definitions for "total order". To address this issue, since definition 2.16 provides what appears to be the intended version of "total order," and the type of order being defined in definition 2.24 (a strict order which is also connected) was already defined under definition 2.22 as a "strict linear order", the following changes were made (including some minor typo fixes and cosmetic changes for section consistency):

Definition 2.24: Removed. The first part of this definition conflicted with definition 2.16, and the second part of this definition was already defined under definition 2.22.

Definition 2.16: Added the label \ollabel{def:linearorder} to this definition and replaced all references to the label \ollabel{def:strictlinearorder}, from definition 2.24, with this label.

Definition 2.22: Added "strict total order" as an appropriate alternative name for "strict linear order" to mirror the approach taken in definition 2.16. However, throughout this section this type of order is still only referred to as a "strict linear order" rather than a "strict total order". This approach was taken to be consistent with the usage of definition 2.16 within the section ("linear order" is preferred over the alternative "total order"). Moved \ollabel{def:strictlinearorder} from the now removed definition 2.24 to this definition.

Proposition 2.25: Modified to refer to "strict linear order" rather than "total" based on the definition fixes. Several typos with "Id_X" were changed to "Id_A". Replaced occurrences of "antisymmetric" with "anti-symmetric" for consistency with the hyphenated spelling of the anti-symmetry definition given in definition 2.6 (but it should be noted that "antisymmetry" also appears elsewhere in the complete text and may be the preferred spelling).

Proposition 2.26: Modified to refer to "strict linear order" rather than "total" based on the definition fixes. The set "X" was changed to "A" for consistency with the rest of the section.

Example 2.27: Removed. This example was incompatible with definition 2.16. Additionally, a similar example, which is consistent with definition 2.16, is already given in example 2.23.

Proposition 2.28: Modified to clarify the use of "total" based on the definition fixes. Label \ollabel{prop:extensionality-totalorders} was changed to \ollabel{prop:extensionality-strictlinearorders} to be consistent with the definition fixes. Added a period to end the proposition sentence. Changed the sentence directly prior to this proposition to properly refer to "strict linear orders" rather than "total orders".

Section 5.6 footnote: This footnote was changed to refer to \ollabel{def:linearorder} (definition 2.16) instead of \ollabel{def:strictlinearorder}. The recalled definition of total order given in this footnote was also missing the anti-symmetric requirement for total orders, as defined in definition 2.16. This requirement is now added. Changed "total ordering" to "total order", for clarity, since the former had not been explicitly defined.